Move ListFilter into filtering.rs
